requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI Integration  
anthropic==0.18.1

# PDF Processing
pdfplumber==0.10.3

# HTTP Client
httpx==0.25.2

# Data Validation - Using older version for Python 3.9 compatibility
pydantic==2.4.2
pydantic-settings==2.0.3

# Environment Management
python-dotenv==1.0.0

# Configuration
pyyaml==6.0.1

# Storage
aiofiles==23.2.1
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2

# Development Tools (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6